// src/app/modules/student/student.model.ts
import { Schema, model, connect } from 'mongoose';
import {
  FullName,
  Guardian,
  LocalGuardian,
  Student,
} from './student.interface';

const fullNameSchema = new Schema<FullName>({
  firstName: { type: String, required: true },
  middleName: { type: String },
  lastName: { type: String, required: true },
});

const gurdianSchema = new Schema<Guardian>({
  fatherName: { type: String, required: true },
  fatherOccupation: { type: String, required: true },
  fatherContactNo: { type: String, required: true },
  motherName: { type: String, required: true },
  motherOccupation: { type: String, required: true },
  motherContactNo: { type: String, required: true },
});

const localGuardianSchema = new Schema<LocalGuardian>({
  name: { type: String, required: true },
  occupation: { type: String, required: true },
  contactNo: { type: String, required: true },
  address: { type: String, required: true },
});

const studentSchema = new Schema<Student>({
  id: { type: String },
  name: fullNameSchema,
  gender: ['male', 'female'],
  dateOfBirth: { type: String },
  contactNo: { type: String, required: true },
  emergencyContactNo: { type: String, required: true },
  bloodGroup: ['A+', 'A-', 'B+', 'B-', 'AB+', 'AB-', 'O+', 'O-'],
  presentAddress: { type: String, required: true },
  permanentAddress: { type: String, required: true },
  guardian: gurdianSchema,
  email: { type: String, required: true },
  avatar: { type: String },
  localGuardian: localGuardianSchema,
  profileImage: { type: String },
  isActive: ['active', 'block'],
});

export const StudentModel = model<Student>('Student', studentSchema);
